+++ include/realsense/camera_thread.h
@@ -1,9 +1,12 @@
 #pragma once
 
-#include "miniros/miniros.h"
+#include "realsense/camera_driver.h"
 #include "realsense/shared_variable.h"
 
 #define CAMERA_PERIOD (40000000)
+
+/* 定义在 camera_thread.cpp 中 */
+extern ObjState objState;
 
 void camera_controller();