+++ include/uskin/uskin_thread.h
@@ -1,10 +1,11 @@
 #pragma once
 
-#include "miniros/miniros.h"
 #include "uskin/sensor_driver.h"
 #include "uskin/shared_variable.h"
 
 #define USKIN_PERIOD (80000000)
+
+extern Force force;
 
 void uskin_controller();