+++ include/wsg_controller/shared_variable.h
@@ -2,8 +2,7 @@
 
 #include <vector>
 #include <queue>
-
-#include "miniros/thread_manager.h"
+#include "miniros/miniros.h"
 
 typedef std::vector<double> WSGCMD;