+++ src/master/main.cpp
@@ -1,18 +1,7 @@
-#include "miniros/miniros.h"
-#include "master/wsg_thread.h"
-#include "master/uskin_thread.h"
-#include "ur5e_controller/ur5e_thread.h"
-#include "realsense/camera_thread.h"
+#include "master/interface.h"
 
 /* 线程管理 */
-ThreadManager threadmanager(2);
-/* 夹爪数据 */
-extern WSGConfig wsgConfig;
-WSGCMD cmd;
-/* 传感器数据 */
-extern Force force;
-/* 机械臂数据 */
-extern UrConfig urConfig;
+ThreadManager threadmanager(1);
 
 int main(int argc, char** argv) {
   /* 开启线程 */
@@ -21,7 +10,10 @@ int main(int argc, char** argv) {
   // threadPool.push_back(std::thread(wsg_controller));
   // threadPool.push_back(std::thread(uskin_controller));
   threadPool.push_back(std::thread(ur5e_controller));
-  threadPool.push_back(std::thread(camera_controller));
+  // threadPool.push_back(std::thread(camera_controller));
+
+  std::vector<double> ori_angle = {0, -98.9*deg2rad, 117.8*deg2rad, -108.9*deg2rad, -90*deg2rad, 90*deg2rad};
+  std::vector<double> off_angle = {0, -98.9*deg2rad, 117.8*deg2rad, -108.9*deg2rad, -90*deg2rad, 70*deg2rad};
 
   /* Initial Grasp */
   THETA jointState = urConfig.get_state();
@@ -29,12 +21,10 @@ int main(int argc, char** argv) {
     int key = scanKeyboard();
     switch (key) {
       case 'g':
-        cmd = {50,100};
-        wsgConfig.push(cmd);
+        ur5e::go_to_joint(off_angle, 3);
         break;
       case 'h':
-        cmd = {100,100};
-        wsgConfig.push(cmd);
+        ur5e::go_home(3);
         break;
       case 27: case 'q':
         miniROS::shutdown();